Add User creation page
This is a test to see if Django's generic create view will work for user creation. This is accomplished by importing the User model into views.py and referencing that model in a UserCreateView that inherits from CreateView. Finally, this UserCreateView class is imported into urls.py and added to the urlpatterns list.