Add owner and created_at fields to Review model
This change includes removing the unique constraint for the venue field on the Review model; more than one review can be created per venue. Also, this commit includes a user.is_authenticated test to the dev.html template as an example of how to check if a user is logged and display content accordingly.